requirement.txt
fastapi
python-binance
uvicorn
cachetools